refactor sink serde
